+++ fleche/doc.ml
@@ -132,7 +132,7 @@ module Diags : sig
 
   (** Build simple diagnostic *)
   val make :
-       ?extra:Lang.Diagnostic.Extra.t list
+       ?data:Lang.Diagnostic.Data.t list
     -> Lang.Range.t
     -> Lang.Diagnostic.Severity.t
     -> Pp.t
@@ -140,7 +140,11 @@ module Diags : sig
 
   (** Build advanced diagnostic with AST analysis *)
   val error :
-    range:Lang.Range.t -> msg:Pp.t -> ast:Node.Ast.t -> Lang.Diagnostic.t
+       lines:string array
+    -> range:Lang.Range.t
+    -> msg:Pp.t
+    -> ast:Node.Ast.t
+    -> Lang.Diagnostic.t
 
   (** [of_messages drange msgs] process feedback messages, and convert some to
       diagnostics based on user-config. Default range [drange] is used for
@@ -152,20 +156,30 @@ module Diags : sig
 end = struct
   let err = Lang.Diagnostic.Severity.error
 
-  let make ?extra range severity message =
-    Lang.Diagnostic.{ range; severity; message; extra }
+  let make ?data range severity message =
+    Lang.Diagnostic.{ range; severity; message; data }
 
   (* ast-dependent error diagnostic generation *)
-  let extra_diagnostics_of_ast ast =
+  let extra_diagnostics_of_ast ~lines ast =
+    let stm_range = ast.Node.Ast.v |> Coq.Ast.loc |> Option.get in
+    let stm_range = Coq.Utils.to_range ~lines stm_range in
+    let stm_range = Lang.Diagnostic.Data.SentenceRange stm_range in
     match Coq.Ast.Require.extract ast.Node.Ast.v with
     | Some { Coq.Ast.Require.from; mods; _ } ->
       let refs = List.map fst mods in
-      Some [ Lang.Diagnostic.Extra.FailedRequire { prefix = from; refs } ]
-    | _ -> None
+      Some
+        [ stm_range
+        ; Lang.Diagnostic.Data.FailedRequire { prefix = from; refs }
+        ]
+    | _ -> Some [ stm_range ]
 
-  let error ~range ~msg ~ast =
-    let extra = extra_diagnostics_of_ast ast in
-    make ?extra range Lang.Diagnostic.Severity.error msg
+  let extra_diagnostics_of_ast ~lines ast =
+    if !Config.v.send_diags_extra_data then extra_diagnostics_of_ast ~lines ast
+    else None
+
+  let error ~lines ~range ~msg ~ast =
+    let data = extra_diagnostics_of_ast ~lines ast in
+    make ?data range Lang.Diagnostic.Severity.error msg
 
   let of_feed ~drange (range, severity, message) =
     let range = Option.default drange range in
@@ -736,7 +750,7 @@ let strategy_of_coq_err ~node ~state ~last_tok = function
   | Coq.Protect.Error.Anomaly _ -> Stop (Failed last_tok, node)
   | User _ -> Continue { state; last_tok; node }
 
-let node_of_coq_result ~token ~doc ~range ~ast ~st ~parsing_diags
+let node_of_coq_result ~lines ~token ~doc ~range ~ast ~st ~parsing_diags
     ~parsing_feedback ~feedback ~info last_tok res =
   match res with
   | Ok state ->
@@ -748,7 +762,7 @@ let node_of_coq_result ~token ~doc ~range ~ast ~st ~parsing_diags
   | Error (Coq.Protect.Error.Anomaly (err_range, msg) as coq_err)
   | Error (User (err_range, msg) as coq_err) ->
     let err_range = Option.default range err_range in
-    let err_diags = [ Diags.error ~range:err_range ~msg ~ast ] in
+    let err_diags = [ Diags.error ~lines ~range:err_range ~msg ~ast ] in
     let contents, nodes = (doc.contents, doc.nodes) in
     let context =
       Recovery_context.make ~contents ~last_tok ~nodes ~ast:ast.v ()
@@ -805,8 +819,8 @@ let document_action ~token ~st ~parsing_diags ~parsing_feedback ~parsing_time
          this point then, hence the new last valid token last_tok_new *)
       let last_tok_new = Coq.Parsing.Parsable.loc doc_handle in
       let last_tok_new = Coq.Utils.to_range ~lines last_tok_new in
-      node_of_coq_result ~token ~doc ~range:ast_range ~ast ~st ~parsing_diags
-        ~parsing_feedback ~feedback ~info last_tok_new res)
+      node_of_coq_result ~lines ~token ~doc ~range:ast_range ~ast ~st
+        ~parsing_diags ~parsing_feedback ~feedback ~info last_tok_new res)
 
 module Target = struct
   type t =