+++ lang/diagnostic.mli
@@ -4,8 +4,9 @@
 (* Written by: Emilio J. Gallego Arias                                  *)
 (************************************************************************)
 
-module Extra : sig
+module Data : sig
   type t =
+    | SentenceRange of Range.t
     | FailedRequire of
         { prefix : Libnames.qualid option
         ; refs : Libnames.qualid list
@@ -28,7 +29,7 @@ type t =
   { range : Range.t
   ; severity : Severity.t
   ; message : Pp.t
-  ; extra : Extra.t list option
+  ; data : Data.t list option [@default None]
   }
 
 val is_error : t -> bool